requirements.txt
aiohttp==1.0.5
async-timeout==1.1.0
boto==2.48.0
cffi==1.9.1
chardet==2.3.0
click==6.7
cryptography==1.7.2
discord.py==0.16.6
Django==1.10.5
Flask==0.12
google-api-python-client==1.6.2
gspread==0.6.2
gunicorn==19.6.0
httplib2==0.10.3
idna==2.2
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
multidict==2.1.4
oauth2client==4.0.0
olefile==0.44
Pillow==4.1.1
praw==5.2.0
prawcore==0.12.0
pyasn1==0.2.2
pyasn1-modules==0.0.8
pycparser==2.17
pygsheets==1.0.0
PyNaCl==1.0.1
pyOpenSSL==16.2.0
pytz==2017.3
PyYAML==3.12
requests==2.13.0
rsa==3.4.2
simplejson==3.11.1
six==1.10.0
update-checker==0.16
uritemplate==3.0.0
virtualenv==15.1.0
websockets==3.2
Werkzeug==0.11.15
